// web/src/app/generated/model/user.ts
/**
 * Modul450 Library API
 * Simple library management system for renting books
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface User { 
    /**
     * Unique identifier for the user
     */
    id: string;
    /**
     * Full name of the user
     */
    name: string;
    /**
     * Email address of the user
     */
    email: string;
    /**
     * Whether the user is authenticated
     */
    authenticated: boolean;
}